add : EditInfo component
