+++ src/assets/scss/EditInfo.module.scss
@@ -1,0 +1,28 @@
+$main-color: #19ce60;
+.contents {
+  .content_box {
+    margin: 30px;
+  }
+  border-bottom: solid 1px #d9d9d9;
+}
+
+.row {
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  font-size: 16px;
+  margin: 10px;
+  .title {
+    width: 150px;
+  }
+  .input_content {
+    width: 600px;
+    height: 20px;
+    font-size: 16px;
+    border: 1px solid #dddddd;
+    padding: 5px;
+  }
+  .content {
+    padding: 5px;
+  }
+}
