+++ src/components/index.ts
@@ -1,0 +1,10 @@
+import BasicButton from "./BasicButton";
+import ContentBox from "./ContentBox";
+import ContentTop from "./ContentTop";
+import EditInfo from "./EditInfo";
+import MyInfo from "./MyInfo";
+import NavBar from "./NavBar";
+import SideBar from "./SideBar";
+import TabContent from "./TabContent";
+
+export { BasicButton, ContentBox, ContentTop, EditInfo, MyInfo, NavBar, SideBar, TabContent };
