+++ src/pages/CompanyPage.tsx
@@ -1,7 +1,6 @@
-import "../assets/scss/Company.scss";
-import ContentTop from "../components/ContentTop";
+import style from "../assets/scss/Company.module.scss";
+import { ContentTop, EditInfo } from "../components";
 import { companyInfoData } from "../assets/data/MockDatas";
-import { useState } from "react";
 
 const companyInfo = [
   { display_name: "사업자번호", data: companyInfoData.business_number, editable: false },
@@ -22,47 +21,16 @@ const CompanyPage = () => {
       <div className="go_content">
         <ContentTop titleName={"회사 관리"} path={"/"} />
       </div>
-      <div className="contents">
-        <div className="content_box">
-          {companyInfo.map((info, index) => (
-            <CompanyRow
-              key={index}
-              displayName={info.display_name}
-              data={info.data}
-              editable={info.editable}
-            />
-          ))}
+      <div className={style.contents}>
+        <div className={style.content_box}>
+          <EditInfo info={companyInfo} />
         </div>
       </div>
-      <div className="footer">
-        <button className="btn_major">저장</button>
-        <button className="btn_minor">취소</button>
+      <div className={style.footer}>
+        <button className={style.btn_major}>저장</button>
+        <button className={style.btn_minor}>취소</button>
       </div>
     </>
-  );
-};
-
-interface CompanyRowProps {
-  displayName: string;
-  data: string;
-  editable: boolean;
-}
-const CompanyRow = (props: CompanyRowProps) => {
-  const [value, setValue] = useState(props.data);
-
-  const onChange = (e: React.ChangeEvent<HTMLInputElement>): void => {
-    setValue(e.target.value);
-  };
-
-  return (
-    <div className="row">
-      <div className="title">{props.displayName}</div>
-      {props.editable ? (
-        <input className="input_content" value={value} onChange={onChange} />
-      ) : (
-        <span className="content">{props.data}</span>
-      )}
-    </div>
   );
 };
